5.2.2 Metaprogramming the trees
Following this part of the book, but in Python, proved to be somewhat difficult. For a start, it seems like in Python it would have been easier to use pattern matching instead of the "visitor" pattern. And in python there are better ways to generate the code rather than using strings, but for the sake of accuracy, I decided to follow the path that has been taken by the book.